# book-network-ui/src/app/services/models/book-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookRequest {
  author_name: string;
  id?: number;
  isbn: string;
  shareable?: boolean;
  synopsis: string;
  title: string;
  bookCover?: String;  // Add the bookCover property (Base64 string or Blob)

}