# book-network-ui/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Emna","description":"OpenApi documentation for Spring security","termsOfService":"Terms of service","contact":{"name":"Emna","url":"http://aliboucoding.com/courses","email":"emna.haddar@supcom.tn"},"license":{"name":"Licence name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8088/api/v1","description":"Local ENV"},{"url":"https://aliboucoding.com/courses","description":"PROD ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/feedbacks":{"post":{"tags":["Feedback"],"operationId":"saveFeedback","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/books":{"get":{"tags":["Book"],"operationId":"findAllBooks","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBookResponse"}}}}}},"post":{"tags":["Book"],"operationId":"saveBook","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/books/cover/{book-id}":{"post":{"tags":["Book"],"operationId":"uploadBookCoverPicture","parameters":[{"name":"book-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/books/borrow/book-id":{"post":{"tags":["Book"],"operationId":"borrowBook","parameters":[],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/books/shareable/{book-id}":{"patch":{"tags":["Book"],"operationId":"updateShareableStatus","parameters":[{"name":"book-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/books/borrow/return/{book-id}":{"patch":{"tags":["Book"],"operationId":"returnBorrowBook","parameters":[{"name":"book-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/books/borrow/return/approve/{book-id}":{"patch":{"tags":["Book"],"operationId":"approveReturnBorrowBook","parameters":[{"name":"book-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/books/archived/{book-id}":{"patch":{"tags":["Book"],"operationId":"updateArchivedStatus","parameters":[{"name":"book-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/feedbacks/book/{book-id}":{"get":{"tags":["Feedback"],"operationId":"findAllFeedbacksByBook","parameters":[{"name":"book-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseFeedbackResponse"}}}}}}},"/books/{book-id}":{"get":{"tags":["Book"],"operationId":"findBookById","parameters":[{"name":"book-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookResponse"}}}}}}},"/books/returned":{"get":{"tags":["Book"],"operationId":"findAllReturnedBooks","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBorrowedBookResponse"}}}}}}},"/books/owner":{"get":{"tags":["Book"],"operationId":"findAllBooksByOwner","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBookResponse"}}}}}}},"/books/borrowed":{"get":{"tags":["Book"],"operationId":"findAllBorrowedBooks","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBorrowedBookResponse"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"FeedbackRequest":{"required":["bookId","comment"],"type":"object","properties":{"note":{"maximum":5,"minimum":0,"type":"number","format":"double"},"comment":{"type":"string"},"bookId":{"type":"integer","format":"int32"}}},"BookRequest":{"required":["authorName","isbn","synopsis","title"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"authorName":{"type":"string"},"isbn":{"type":"string"},"synopsis":{"type":"string"},"shareable":{"type":"boolean"}}},"RegistrationRequest":{"required":["email","firstname","lastname","password"],"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"FeedbackResponse":{"type":"object","properties":{"note":{"type":"number","format":"double"},"comment":{"type":"string"},"ownFeedback":{"type":"boolean"}}},"PageResponseFeedbackResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/FeedbackResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"BookResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"authorName":{"type":"string"},"isbn":{"type":"string"},"synopsis":{"type":"string"},"owner":{"type":"string"},"cover":{"type":"array","items":{"type":"string","format":"byte"}},"rate":{"type":"number","format":"double"},"archived":{"type":"boolean"},"shareable":{"type":"boolean"}}},"PageResponseBookResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/BookResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"BorrowedBookResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"authorName":{"type":"string"},"isbn":{"type":"string"},"rate":{"type":"number","format":"double"},"returned":{"type":"boolean"},"returnApproved":{"type":"boolean"}}},"PageResponseBorrowedBookResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/BorrowedBookResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth descritption","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}