# book-network/src/main/java/com/example/book/book/BookRequest.java
package com.example.book.book;

import com.fasterxml.jackson.annotation.JsonProperty;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;

public record BookRequest(
        Integer id,
        @NotNull(message = "100")
        @NotEmpty(message = "100")
        String title,
        @JsonProperty("author_name")

        //@NotNull(message = "101")
        //@NotEmpty(message = "101")
        String author_name,
        String bookCover,
        @NotNull(message = "102")
        @NotEmpty(message = "102")
        String isbn,
        @NotNull(message = "103")
        @NotEmpty(message = "103")
        String synopsis,
        boolean shareable
) {


        public String getBookCover() {
                return bookCover;
        }
}